+++ src/PokeAByte.Domain/Interfaces/MapperContent.cs
@@ -1,0 +1,9 @@
+﻿namespace PokeAByte.Domain.Interfaces;
+
+public enum MapperFileType : byte
+{
+    Official,
+    Local
+}
+
+public record MapperContent(string Path, string Xml, string? ScriptPath, string? ScriptRoot);