+++ src/PokeAByte.Web/Controllers/DriverEndpoints.cs
@@ -1,10 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using PokeAByte.Domain.Interfaces;
+using PokeAByte.Web.Models;
 using PokeAByte.Web.Services.Mapper;
 
 namespace PokeAByte.Web.Controllers;
-
-public record UpdateMemoryModel(uint Address, byte[] Bytes);
 
 public static class DriverEndpoints
 {
@@ -12,6 +11,11 @@ public static class DriverEndpoints
     {
         app.MapPut(
             "/diver/memory",
+            async (IStaticMemoryDriver driver, [FromBody] UpdateMemoryModel model)
+                => await driver.WriteBytes(model.Address, model.Bytes)
+        );
+        app.MapPut(
+            "/driver/memory",
             async (IStaticMemoryDriver driver, [FromBody] UpdateMemoryModel model)
                 => await driver.WriteBytes(model.Address, model.Bytes)
         );