+++ src/PokeAByte.Web/Models/GlossaryItemModel.cs
@@ -1,3 +1,3 @@
-﻿namespace PokeAByte.Domain.Models.Properties;
+﻿namespace PokeAByte.Web.Models;
 
 public record GlossaryItemModel(ulong Key, object? Value);